Test for existing Years in Dictionary
The addition of the model year lookup dictionary had abismal results on my completion percentage. I suspected this was because of a glitch in creating the dictionary or the lookup syntax. I tested all of these and have come to the hypothesis that there are insufficient model year entries in the key file.